requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
requests==2.32.3
pygithub==2.3.0
python-jose==3.3.0
pydantic==2.9.2
